\name{alon}
\docType{data}
\alias{alon}
\title{
Data from the microarray experiment done by Alon et al. (1999)
}
\description{
The data set contains the gene expression data origianlly analized in Alon et al. (1999). 62 samples (40 tumor samples, 22 normal samples) from colon-cancer patients were analyzed with an Affymetrix oligonucleotide Hum6000 array. The binary variable \code{y} is used to indicate a normal sample (\code{y = 0}) or a tumor sample (\code{y = 1}).

Two thousand out of around 6500 genes were selected based on the confidence in the measured expression levels (for details refer to publication).
}
\usage{data(alon)}
\references{
Alon U., Barkai N., Notterman D.A., Gish K., Ybarra S., Mack D. and Levine A.J. (1999) <doi:10.1073/pnas.96.12.6745>
\emph{Broad patterns of gene expression revealed by clustering analysis of tumor and normal colon tissue probed by oligonucleotide arrays.} \emph{Proc. Natl. Acad. Sci. USA} \bold{96}, 6745-6750
}
\keyword{datasets}
