\name{dglars-internal}
\title{Internal dglars functions}
\alias{dglars_pc}
\alias{dglars_ccd}
\alias{make_dglars}
\alias{make_coef}
\alias{make_ru_pc}
\alias{make_action}
\alias{make_summary_table}
\alias{setDiff}
\alias{cvdglars_pc}
\alias{cvdglars_ccd}
\alias{make_cvdglars}
\description{Internal dglars functions}
\author{Luigi Augugliaro}
\details{These are not intended for use by users}
\keyword{internal}