\name{duke}
\docType{data}
\alias{duke}
\title{
Duke breast cancer microarray experiment
}
\description{
This data set details microarray experiment for 44 breast cancer patients. The binary variable \code{Status} is used to classify the patients into estrogen receptor-positive (\code{Status = 0}) and estrogen receptor-negative (\code{Status = 1}). The other variables  contain the expression level of the considered genes.
}
\usage{data(duke)}
\references{
M. West, C. Blanchette, H. Dressman, E. Huang, S. Ishida, R. Spang, H. Zuzan, J.A. Olson, Jr., J.R. Marks and Joseph R. Nevins (2001)
\emph{Predicting the clinical status of human breast cancer by using gene expression profiles}, \emph{Proceedings of the National Accademy of Sciences of the USA}, Vol 98(20), 11462-11467.
}
\keyword{datasets}
