\name{dglars-internal}
\title{Internal dglars Functions}
\alias{d2th_dmu2_mk}
\alias{d2mu_de2_mk}
\alias{dglars_pc}
\alias{dglars_ccd}
\alias{make_dglars}
\alias{make_coef}
\alias{make_ru}
\alias{make_action}
\alias{make_summary_table}
\alias{setDiff}
\alias{cvdglars_pc}
\alias{cvdglars_ccd}
\alias{make_cvdglars}
\description{Internal dglars functions.}
\author{Luigi Augugliaro}
\details{These functions are not intended for use by users.}
\keyword{internal}
